DXFxtctrV1
So far with the extractor, the GUI is set up, and a certain level of extraction. Only the name of entities can be extracted as of now, I'm yet to properly understand the ezdxf documentation and how attributes of the entities can be extracted. And so the information can be seen in it's final data frame, it can be saved to a CSV file to see how the code is progressing in extraction capacity.